webtronics/apps/likes/serializers.py
from rest_framework import serializers
from .models import Like
from posts.models import Post

class LikeSerializer(serializers.ModelSerializer):
    id = serializers.IntegerField(read_only=True)
    user = serializers.HiddenField(default=serializers.CurrentUserDefault())
    post = serializers.PrimaryKeyRelatedField(queryset=Post.objects.all())
    class Meta:
        model = Like
        fields = '__all__'